pyproject.toml
[project]
name = "jpcv"
version = "0.1.0"
description = "職務経歴書 PDF Generator - Generate Japanese work history PDFs from YAML"
# readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "reportlab>=4.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.scripts]
jpcv = "jpcv.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jpcv"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
